spring security

+++ journal-app-edigest-springboot/journal-App/src/main/java/com/springbootedigest/journal/App/service/JournalEntryService.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -23,10 +24,10 @@ public class JournalEntryService {
     @Transactional
     public void saveEntry(JournalEntry journalEntry, String username) {
         User user = userService.findByUserName(username);
-
+        journalEntry.setDate(LocalDateTime.now());
         JournalEntry save = journalEntryRepository.save(journalEntry);
         user.getJournalEntries().add(save);
-        userService.saveEntry(user);
+        userService.saveUser(user);
     }
 
     public List<JournalEntry> findAll() {
@@ -37,14 +38,28 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String username) {
-        User user = userService.findByUserName(username);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String username) {
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(username);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        } catch (Exception e) {
+            throw new RuntimeException("An error occurred");
+        }
+
+        return removed;
     }
 
     public void saveEntry(JournalEntry oldJournal) {
         journalEntryRepository.save(oldJournal);
     }
+
+//    public List<JournalEntry> findByUsername(String username) {
+//        return null;
+//    }
 }
